Add files via upload
Hi Guys, so I uploaded the script plot3D.py . Its kind of a template, you have to add how to load the dataset. The important part is the last one. One issue with this 3d plot is more complicate to color by class, I think seaborn does not support 3D plots, so "hue" parameter is not available.
Tot snell